Fix code scanning alert #2: Multiplication result converted to larger type

+++ grm.cpp
@@ -256,7 +256,7 @@ void gcta::make_grm(bool grm_xchr_flag, bool output_bin, int grm_mtd)
 	    if(grm_mtd==1){
             for(k=0, denom=0.0; k<_include.size(); k++){
                 if(X[i][k]<1e5){
-                    A[i][i]+=X[i][k]*X[i][k];
+                    A[i][i]+=static_cast<double>(X[i][k]) * X[i][k];
                     denom+=d[k];
                     A_N[i][i]++;
                 }
